webapi/tct-application-tizen-tests/application/ApplicationManager_launchAppControl_with_appControl_operation_mime.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Junghyuk Park <junghyuk.park@samsung.com>


-->

<html>

<head>
<title>ApplicationManager_launchAppControl_with_appControl_operation_mime</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/app_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: ApplicationManager_launchAppControl_with_appControl_operation_mime
//==== LABEL Check using launchAppControl method (with operation and MIME type of ApplicationControl) in ApplicationManager interface
//==== SPEC Tizen Web API:Application:Application:ApplicationManager:launchAppControl M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/application.html
//==== PRIORITY P1
//==== ONLOAD_DELAY 30
//==== TEST_CRITERIA MMINA MAST

setup({timeout: 30000});

var t = async_test(document.title, {timeout: 30000}),
    appcontrol;

setup_launch(t, TCT_APPCONTROL_APPID, function () {
    appcontrol = new tizen.ApplicationControl(
        TCT_APPCONTROL_LAUNCH_APPCONTROL_OPERATION,
        null,
        TCT_APPCONTROL_LAUNCH_APPCONTROL_MIME
    );

    tizen.application.launchAppControl(appcontrol);

    assert_launch(t, TCT_APPCONTROL_APPID, function () {
        t.done();
    });
});

</script>
</body>
</html>
